[trivial] internal: refactor figma-tasks, update to dynamically generated pngs (#1481)
* [trivial] internal: refactor figma-tasks, update to dynamically generate pngs

* fix icons sync

* fix glyphMap generation for icons

* undo changes to common's package.json

* svgJs generation and align versioning to old versions

* temporarily commit illustration styles generated content

* update sync-illustrations

* before macos update

* run yarn dedupe

* undo common package.json changes

* align all figma-tasks output

* remove confusing comments

* improve comments on sync-illustrations

* fix script-utils build

* improve comment for requestType

* fix lint error

* cleanup comments and re-use of utils

* align icons and illustrations comments

* fix sync-icons

* ensure icon manifest.json is in generated directory

+++ packages/figma-tasks/src/helpers/font/fontOutputGenerator.ts
@@ -1,7 +1,7 @@
 import fs from 'node:fs';
 import path from 'node:path';
 import { Task } from '@cbhq/mono-tasks';
-import { getAbsolutePath, writePrettyFile } from '@cbhq/script-utils';
+import { existsOrCreateDir, getAbsolutePath, writePrettyFile } from '@cbhq/script-utils';
 
 import { getFontNameAndHash } from './getFontNameAndHash';
 import { getGlyphMapContent } from './getGlyphMapContent';
@@ -39,11 +39,12 @@ export function fontOutputGenerator({
       const fontNameWithExt = `${fontName}.${generatedFontFormat}`;
       const fontFilePath = path.join(fontOutputDir, fontNameWithExt);
       const fontFileDir = path.dirname(fontFilePath);
-      if (hash) {
+      if (hash && fs.existsSync(fontFileDir)) {
         fs.rmSync(fontFileDir, { recursive: true });
         fs.mkdirSync(fontFileDir, { recursive: true });
       }
 
+      await existsOrCreateDir(fontFilePath);
       promises.push(fs.promises.writeFile(fontFilePath, fontContents));
 
       if (generatedGlyphMapFile) {
@@ -70,6 +71,8 @@ export function fontOutputGenerator({
         }`;
         promises.push(writePrettyFile(cssFilePath, cssContent));
       }
+
+      await Promise.all(promises);
     }
   };
 }
